requirements-global.txt
# base requirements
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
tqdm==4.66.4
PyYAML==6.0.1
huggingface-hub==0.23.3
scipy==1.13.1; sys_platform != 'win32'
matplotlib==3.9.0

# pytorch
accelerate==0.30.1
safetensors==0.4.3
tensorboard==2.17.0
pytorch-lightning==2.2.5

# stable diffusion
-e git+https://github.com/huggingface/diffusers.git@e45c25d#egg=diffusers
transformers==4.42.3
omegaconf==2.3.0 # needed to load stable diffusion from single ckpt files
invisible-watermark==0.2.0 # needed for the SDXL pipeline

# other models
pooch==1.8.1
open-clip-torch==2.24.0

# data loader
-e git+https://github.com/Nerogar/mgds.git@fa78a18#egg=mgds

# optimizers
dadaptation==3.2 # dadaptation optimizers
lion-pytorch==0.1.4 # lion optimizer
prodigyopt==1.0 # prodigy optimizer
-e git+https://github.com/Sayat1/schedule_free#egg=schedulefree
#schedulefree==1.2.7 # schedule-free optimizers
-U pytorch_optimizer # pytorch optimizers

# Profiling
scalene==1.5.41

# ui
customtkinter==5.2.2
